+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased] - YYYY-MM-DD
 
+## [0.3.0] - 2019-12-20
+
 ### Added
 
 - Forward alerts by custom chat ID based on alert labels.
@@ -42,7 +44,8 @@
 - Alertmanager compatible webhook API.
 - Models and forwarding domain service.
 
-[unreleased]: https://github.com/slok/alertgram/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/slok/alertgram/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/slok/alertgram/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/slok/alertgram/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/slok/alertgram/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/slok/alertgram/releases/tag/v0.1.0